refactor(wardens): extract verdict-store capsule from gate engine (LIA-306)
Step 2 of decomposing the ~4000-line codex_warden_hooks.py. Moves the 13
verdict-store I/O functions (_verdicts_path, _read/_write/_clear_verdict,
_last_verdict*, _audit/_bypass_log_path, _write_bypass_log,
record_script_verdict) verbatim into a new warden_hooks/verdict_store.py.

Unlike step-1's pure leaves, this cluster transitively depends on entry-module
helpers that tests monkeypatch (_claude_marker_dir, _git) plus _write_atomic,
_debug, _marker_dir_for_worktree, MARKER_NAMES. To keep those monkeypatches
effective without re-importing the entry on the hot hook path, the entry injects
itself via bind_entry(sys.modules[__name__]); the capsule resolves entry-owned
helpers through that live reference at call time (9 call sites, 6 symbols).
Intra-cluster calls stay direct. All symbols re-exported so hooks.<name> and
codex_warden.py's import surface are unchanged.

Also drops the dead PurePosixPath import.

Entry: -119 net (4119 -> 4000). No behavior change.
